maint: enable linting step in CI  (#92)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

- Closes #31 

## Short description of the changes

built off of #85 , please review that one first! Only the last 3 commits
in this PR are finalizing turning on the linting in CI.
## How to verify that this has the expected result

In CircleCI, the linting step should run.