maint: add coverage report html to circleci artifacts (#93)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

- Closes #84 

## Short description of the changes
Improves maintainability! Uses coverage.py to run our test suite and
compare it to our src directory to find un-covered features. Uploads the
report as an interactive html to CircleCI artifacts.

## How to verify that this has the expected result
In CircleCI, under a test workflow, go to the Artifacts tab and find the
[htmlcov/index.html](https://output.circle-artifacts.com/output/job/ff865a0e-797d-4482-998f-e33a1ef90b7d/artifacts/0/htmlcov/index.html)
for code coverage. It's interactive to find different functions in files
we haven't covered in testing!