maint: Add nightly tests job (#146)
## Which problem is this PR solving?
Adds nightly test jobs and upgrades ubuntu and the collector version for
smoke tests. It seems like this also fixes the failed smoke test issues
we're seeing on main.

- Closes #145 

## Short description of the changes
- Added `nightly` workflow to CircleCI config
- Upgraded Ubuntu for smoke tests
- Upgraded collector version for smoke tests

## How to verify that this has the expected result
- [Nightly build runs
tests](https://app.circleci.com/pipelines/github/honeycombio/honeycomb-opentelemetry-python/388/workflows/6b3f9ebe-f150-4abb-a524-238d214668e3)
- Not sure how to test the cron job will work exactly before we merge
this in, but it is the same as on our other repos so I'm fairly
confident it will be the same.