main: Use less restrictive package version for instrumentation dependency (#196)
## Which problem is this PR solving?
The dependency on `opentelemetry-instrumentation` is overly restrictive
and doesn't allow later minor versions to be used. For example, it's set
as `~0.43b0` but doesn't allow the latest releases `0.44b0` or `0.45b0`.

This is because the
[tilde](https://python-poetry.org/docs/dependency-specification/#tilde-requirements)
`~` version constraint doesn't allow versions that change within the
same semver point specified. The version we set is `0.43b0`, which means
it can't go higher than minor release `0.43b0` and subsequent packages
are minor releases. It would allow patch version of the same minor if
they were published, eg `0.43b0.x`.

- Closes #195 

## Short description of the changes
- Update `opentelemetry-instrumentation` package version constraint to
use the
[caret](https://python-poetry.org/docs/dependency-specification/#caret-requirements)
`^` prefix instead, which is equivalent to >= 0.43b0 and < 1.0

## How to verify that this has the expected result
Later version of opentelemetry-instrumentation package can be used with
the distro.